third_party/blink/web_tests/fast/css/getComputedStyle/computed-style-redistribution-expected.txt

Check that getComputedStyle causes a shadow re-distribution when necessary.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(d1).color is ''
PASS getComputedStyle(d2).color is green
PASS getComputedStyle(d1).color is green
PASS getComputedStyle(d2).color is ''
PASS successfullyParsed is true

TEST COMPLETE

A
